Added -r
Added a -r option to rebuild the rss feed w/o recording. This is useful
if you have made changes to the files on disk (such as deleting old
recordings).
Moved the "recording to" so it only appears when recording.
Added a log message to indicate when ID3 tagging starts.